MFC systrace intercept_translate() problem resulting in problems with systrace and some *at functions, most noticable building ports using gtar under systrace.
Date: 2012/03/02 17:23:08
Author: matthew
Branch: HEAD
Tag: (none)
Log:
Fix handling of 'offend' (OFFset from the END of the arg list)
arguments in systrace(1).  In intercept_translate(), argsize is
actually the number of *bytes* taken up by the arguments, not the
number of arguments.

ok ajacoutot, sthen, dcoppa

Members:
        intercept-translate.c:1.14->1.15

+++ usr.bin/less/lsystem.c
@@ -138,7 +138,7 @@ lsystem(cmd, donemsg)
 			char *esccmd = shell_quote(cmd);
 			if (esccmd != NULL)
 			{
-				int len = strlen(shell) + strlen(esccmd) + 5;
+				size_t len = strlen(shell) + strlen(esccmd) + 5;
 				p = (char *) ecalloc(len, sizeof(char));
 				SNPRINTF3(p, len, "%s %s %s", shell, shell_coption(), esccmd);
 				free(esccmd);
